tn-service/services/worker/requirements.txt
redis==5.0.8
requests==2.32.3
openai==1.61.0
pillow==10.4.0
psycopg[binary]==3.2.1
